DBusMessageIter: Increase size for 128-bit pointers
DBusMessageIter is an opaque structure that should/cannot directly be used by
users of libdbus. Internally it maps to DBusMessageRealIter. In order to allow
users allocations on the stack, DBusMessageIter is defined as user accessible
dummy structure. Static assertions ensure that DBusMessageRealIter fits into
this dummy structure. With 128-bit pointers in Cheri, the alignment changes in
both structures, breaking the assertion.
This commit adapts the dummy structure to reflect 128-bit pointers. It
furthermore removes assertions regarding ABI backward compatibility, since this
cannot be done and isn't required on the new platform.